[ATR-503] refactor: 뉴스레터 구독 api 리팩터링 (#224)
- 뉴스레터를 구독하는 api archive -> subscription으로 파일이동
- 이름 변경 addNewsletter -> subscribeNewsletter
- 기존 하나의 메서드가 3가지일 수행 -> 3개의 메서드가 각 한 가지일 수행
