+++ src/main/java/run/attraction/api/v1/archive/controller/ArchiveController.java
@@ -16,12 +16,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import run.attraction.api.v1.archive.dto.ArticleDTO;
-import run.attraction.api.v1.archive.dto.NewsletterEmail;
 import run.attraction.api.v1.archive.dto.request.UserArticlesRequest;
 import run.attraction.api.v1.archive.dto.response.ApiResponse;
 import run.attraction.api.v1.archive.service.ArchiveService;
-import run.attraction.api.v1.gmail.dto.UserGmailDto;
-import run.attraction.api.v1.gmail.service.GmailService;
 import run.attraction.api.v1.introduction.Newsletter;
 
 
@@ -32,7 +29,6 @@ import run.attraction.api.v1.introduction.Newsletter;
 public class ArchiveController {
 
   private final ArchiveService archiveService;
-  private final GmailService gmailService;
 
   @GetMapping("/{userEmail}/articles")
   public ApiResponse<Page<ArticleDTO>> getUserArticles(@PathVariable String userEmail, @ModelAttribute UserArticlesRequest request) {
@@ -52,14 +48,6 @@ public class ArchiveController {
   @PutMapping("/{userEmail}/article/{articleId}")
   public ApiResponse<Void> saveUserArticleProgress(@PathVariable String userEmail, @PathVariable @NotNull @Min(1) Long articleId, @RequestParam @NotNull int readPercentage) {
     archiveService.saveUserArticleProgress(userEmail, articleId, readPercentage);
-
-    return ApiResponse.from(HttpStatus.OK, "성공", null);
-  }
-
-  @PutMapping("/{userEmail}/subscribe/{newsletterId}")
-  public ApiResponse<Void> addNewsletter(@PathVariable String userEmail, @PathVariable @NotNull @Min(1) Long newsletterId) {
-    final NewsletterEmail newsletterEmail = archiveService.addNewsletter(userEmail, newsletterId);
-    gmailService.applyLabelAndFilter(new UserGmailDto(userEmail, newsletterEmail.email()));
 
     return ApiResponse.from(HttpStatus.OK, "성공", null);
   }
